requirements.txt
torch~=2.5.1
numpy>=1.23.0,<2.0.0
opencv-python~=4.10.0.84
PyYAML~=6.0.2
ultralytics~=8.3.34
sympy~=1.13.1
mediapipe~=0.10.18
pandas~=2.2.3
pydantic~=2.9.2
tqdm~=4.66.3
scikit-learn~=1.5.2
absl-py~=2.1.0
PyQt5~=5.15.11
lapx>=0.5.2